requirements.txt
click==6.6
Pillow==3.2.0
picamera==1.10
RPi.GPIO==0.6.2
